Remove ConfigItem model
