+++ server/lib/config/index.js
@@ -1,5 +1,53 @@
 const nonUi = require('./nonUi')
 const ui = require('./ui')
+const definitions = require('../../resources/configItems')
+
+function makeSave(db) {
+  return function save(key, value) {
+    return new Promise((resolve, reject) => {
+      const definition = definitions.find(definition => definition.key === key)
+
+      if (definition.interface !== 'ui') {
+        return reject(
+          new Error(
+            'Config Item ' +
+              this.key +
+              ' must use ui interface to be saved to db'
+          )
+        )
+      }
+
+      db.config.findOne({ key }).exec(function(err, doc) {
+        if (err) {
+          return reject(err)
+        }
+        if (doc) {
+          doc.value = value
+          doc.modifiedDate = new Date()
+          db.config.update({ _id: doc._id }, doc, {}, function(err) {
+            if (err) {
+              return reject(err)
+            }
+            return resolve(null, doc)
+          })
+        } else {
+          const newConfigValue = {
+            key,
+            value,
+            createdDate: new Date(),
+            modifiedDate: new Date()
+          }
+          db.config.insert(newConfigValue, function(err) {
+            if (err) {
+              return reject(err)
+            }
+            return resolve(null, newConfigValue)
+          })
+        }
+      })
+    })
+  }
+}
 
 /**
  * Gets config helper using all config sources
@@ -25,6 +73,18 @@ exports.getHelper = function getAllConfig(db) {
         }
         return all[key]
       },
+      getItems: () => {
+        return definitions.map(definition => {
+          return Object.assign({}, definition, {
+            effectiveValue: all[definition.key],
+            envValue: envConfig[definition.key],
+            cliValue: cliConfig[definition.key],
+            savedCliValue: savedCliConfig[definition.key],
+            dbValue: uiConfig[definition.key]
+          })
+        })
+      },
+      save: makeSave(db),
       smtpConfigured: () =>
         all.smtpHost &&
         all.smtpUser &&
